Write to temporary filename during download

+++ syncrypt/pipes/io.py
@@ -1,4 +1,5 @@
 import os.path
+import shutil
 
 import aiofiles
 import asyncio
@@ -36,11 +37,12 @@ class FileReader(Source):
 
 class FileWriter(Sink):
     # simple wrapper for aiofiles
-    def __init__(self, filename, create_dirs=False, create_backup=False):
+    def __init__(self, filename, create_dirs=False, create_backup=False, store_temporary=False):
         self.filename = filename
         self.handle = None
         self.create_dirs = create_dirs
         self.create_backup = create_backup
+        self.store_temporary = store_temporary
         super(FileWriter, self).__init__()
 
     @asyncio.coroutine
@@ -49,14 +51,32 @@ class FileWriter(Sink):
             fn = self.filename
             if self.create_dirs and not os.path.exists(os.path.dirname(fn)):
                 os.makedirs(os.path.dirname(fn))
-            if self.create_backup and os.path.exists(fn):
+            if self.create_backup and os.path.exists(fn) and not self.store_temporary:
                 shutil.move(fn, self.get_backup_filename(fn))
+            if self.store_temporary:
+                fn = self.get_temporary_filename(fn)
             self.handle = yield from aiofiles.open(fn, 'wb')
         contents = yield from self.input.read(count)
         yield from self.handle.write(contents)
         return contents
 
+    @asyncio.coroutine
+    def finalize(self):
+        fn = self.filename
+        if self.store_temporary: # we only wrote a temporary filename 
+            if os.path.exists(fn):
+                if self.create_backup:
+                    shutil.move(fn, self.get_backup_filename(fn))
+                else:
+                    os.remove(fn)
+            shutil.move(self.get_temporary_filename(fn), fn)
+
+    def get_temporary_filename(self, filename):
+        # TODO: more elaborate temporary filename composition required
+        return filename + '.sctemp000'
+
     def get_backup_filename(self, filename):
+        # TODO: more elaborate backup filename composition required
         return filename + '.scbackup'
 
     @asyncio.coroutine
